Having unpacking removes the need to store the original dense weights in the python LInear module

Differential Revision: [D34699287](https://our.internmc.facebook.com/intern/diff/D34699287/)

[ghstack-poisoned]

+++ torch/ao/nn/sparse/quantized/dynamic/linear.py
@@ -34,7 +34,7 @@ class Linear(torch.nn.Module):
                                                 scale=1, zero_point=0, dtype=torch.qint8)
         self._packed_params = linear.LinearPackedParams(row_block_size=row_block_size,
                                                         col_block_size=col_block_size,
-                                                        dtype=torch.qint8)
+                                                        dtype=dtype)
         self._packed_params.set_weight_bias(qweight, bias, row_block_size, col_block_size)
 
     def _get_name(self):
